--- a/src/Nethermind/Nethermind.State/VerkleStateTree.cs
+++ b/src/Nethermind/Nethermind.State/VerkleStateTree.cs
@@ -144,5 +144,4 @@ public class VerkleStateTree(IVerkleTreeStore stateStore, ILogManager logManager
 
         return tree;
     }
-
 }