- Commit multiple stem in bulk.
  - Prevent multiple re-hash of parent.
  - Make it easy to parallelize.
- Parallelize commit.
- Parallelize account set.
- Note: my cpu is an AMD R9 7950X with 16 hyper threaded core.

### Before
| Method                         | TreeType | Configuration | Mean          | Error      | StdDev     |
|------------------------------- |--------- |-------------- |--------------:|-----------:|-----------:|
| InsertAndCommitRepeatedlyTimes | MPT      | Balance       |     25.371 ms |  0.4987 ms |  0.6484 ms |
| InsertAndCommitRepeatedlyTimes | MPT      | WriteHeavy    |     20.974 ms |  0.2167 ms |  0.1810 ms |
| InsertAndCommitRepeatedlyTimes | MPT      | ReadHeavy     |      3.114 ms |  0.0346 ms |  0.0324 ms |
| InsertAndCommitRepeatedlyTimes | MPT      | AddressOnly   |     24.475 ms |  0.4765 ms |  0.6361 ms |
| InsertAndCommitRepeatedlyTimes | Verkle   | Balance       |  2,943.736 ms | 39.8139 ms | 37.2420 ms |
| InsertAndCommitRepeatedlyTimes | Verkle   | WriteHeavy    |  2,235.722 ms |  4.2694 ms |  3.9936 ms |
| InsertAndCommitRepeatedlyTimes | Verkle   | ReadHeavy     |    348.186 ms |  5.1596 ms |  4.5739 ms |
| InsertAndCommitRepeatedlyTimes | Verkle   | AddressOnly   | 13,028.443 ms | 71.0158 ms | 66.4282 ms |

### After
| Method                         | TreeType | Configuration | Mean      | Error     | StdDev   |
|------------------------------- |--------- |-------------- |----------:|----------:|---------:|
| InsertAndCommitRepeatedlyTimes | Verkle   | Balance       | 335.46 ms |  6.520 ms | 5.779 ms |
| InsertAndCommitRepeatedlyTimes | Verkle   | WriteHeavy    | 299.38 ms |  4.217 ms | 3.738 ms |
| InsertAndCommitRepeatedlyTimes | Verkle   | ReadHeavy     |  72.18 ms |  1.113 ms | 0.987 ms |
| InsertAndCommitRepeatedlyTimes | Verkle   | AddressOnly   | 775.29 ms | 10.384 ms | 9.205 ms |

## Types of changes

#### What types of changes does your code introduce?

- [ ] Bugfix (a non-breaking change that fixes an issue)
- [ ] New feature (a non-breaking change that adds functionality)
- [ ] Breaking change (a change that causes existing functionality not to work as expected)
- [ ] Optimization
- [ ] Refactoring
- [ ] Documentation update
- [ ] Build-related changes
- [ ] Other: _Description_

## Testing

#### Requires testing

- [ ] Yes
- [ ] No

#### If yes, did you write tests?

- [ ] Yes
- [ ] No

#### Notes on testing

_Optional. Remove if not applicable._

## Documentation

#### Requires documentation update

- [ ] Yes
- [ ] No

_If yes, link the PR to the docs update or the issue with the details labeled `docs`. Remove if not applicable._

#### Requires explanation in Release Notes

- [ ] Yes
- [ ] No

_If yes, fill in the details here. Remove if not applicable._

## Remarks

_Optional. Remove if not applicable._